strip configuration of dhcp (broken) and add tz+pwd vars
remove NET_ADMIN cap and unnecessary ports for dns since
dhcp does not seem to work.
keep only 80/tcp and 53/udp which is needed for dns to work
and for web ui/gui to be available.
parametrize and remove password auth by default.
parametrize and use utc timezone by default.